Fix crash on exit
Forcing the creation of a new OSG GraphicsContext ID on the
CompositeViewer initialization code triggers a crash on exit under
specific circumstances.

The backtrace/Valgrind reported that the GLExtensions object associated
to the second GraphicsContext was getting freed twice.

This change probably breaks something in CompositeViewer, but a crash on
exit is a big nono and fixing it comes first.