Rework how the OpenGL version is requested
Instead of relying on DisplaySettings, now we set the OpenGL version
manually in the GraphicsContext's traits. If context creation fails,
this version is not supported by the driver.

First we attempt to optionally create a 4.3 core context. If that fails,
we create a 4.1 core context. If that fails as well, we crash and warn
the user as that's our minimum required version.

Also purge some useless options from the GraphicsContext traits. Some of
them should probably be configured from the Compositor, but because we
use FBOs intensively, we don't need to mess with the default framebuffer
much.